// src/app/ficha-de-matricula/models/matricula.ts
export class Apoderado {
    matriculaId: number = 0;
    tipoApoderado: string = '';
    rut: string = '';
    nombre: string = '';
    apellidoPaterno: string = '';
    apellidoMaterno: string = '';
    parentescoId: number = 0;
    generoId: number = 0;
    fechaDeNacimiento: string = '';
    paisId: number = 0;
    etniaId: number = 0;
    religionId: number = 0;
    estadoCivilId: number = 0;
    tenenciaDeViviendaId: number = 0;
    calle: string = '';
    numero: number = 0;
    departamento: string = '';
    regionId: number = 0;
    comunaId: number = 0;
    celular: string = '';
    telefonoFijo: string = '';
    correoElectronico: string = '';
    nivelEducacionalId: number = 0;
    profesionId: number = 0;
    lugarDeTrabajo: string = '';
    celularTrabajo: string = '';
    telefonoFijoTrabajo: string = '';
    correoElectronicoTrabajo: string = '';
}